requirements.txt
PyQt5>=5.15
python-docx>=1.1
Pillow>=10.0
qt-material>=2.14
pywin32>=306; platform_system == "Windows"
